[ui] Keep config wizard view in location rather than hash
We used to keep the view state for the config wizard in the URL hash
resulting in URLs like: /setup/wizard#view="morio.node_count"

Those are not very clean URLs so instead I switched to keeping this
state in the URL location. For that to work the following changes were
made:

- Add a wildcard route under /setup that catches all URLs
- Modify the way we set the view by prefixing it with the prefix under
  which the wizard is anchored (/setup/wizard by default) and replacing
  dots with slashes

This means the following change:
- old url: /setup/wizard#view="morio.node_count"
- new url: /setup/wizard/morio/node_count

Which feels a lot cleaner and will make it easier to deep-link.