wip: Work on configuration handling in sam/api
This is a start on working on configuration handling.
We want to allow people to revert to earlier configurations in case they
mess up, so we store a timestamped config file on disk.

Furthermore, config should be backup-safe, so we write an additional
file that holds keys (JWT key and the private key passphrase).
To add another layer of protection ffrom curious eyes, the latter
is written as binary (BSON) so it can't be casually read (but can be
read if you have access to the file system)