wip: Work on letting SAM handle starting of services
Up until this point, we were using docker-compose to start various
containers for a minimal Morio setup. We're not starting to let SAM
handle starting of containers in preparation of it being the central
coordinating component.

The idea is that you start SAM (which we will probably rename to core)
and that it will read your config and start whatever is needed itself.