[ca] Morio can now bootstrap its own CA
So, after the ephemeral stage, Morio will spin up a CA and configure
Traefik to get certificates from the CA via the ACME protocol.

The CA (step-ca) would typically generate it's own root certificate when
initialized. However, that would be it non-deterministic in case we want
to deploy a cluster where we want multiple CAs to have the same root
certificate.

So instead, Morio core will generate the root certificate and
intermediate certificate, sign it, and preconfigure the CA so that no
inistialization is required and it it is ready to go on the first boot.