First refactor to clarify various uses of config
This started out as a couple of improvements, but then become a bit of a
mini-refactor. We were using 'config' in several places to refer to
different things, and that's not very intuitive. So here I started
making efforts to be more explicit. For example:

- `mConf` refers to the high-level Morio configuratino
- `mConf.core` no longer holds things that apply to the entire
  deployment, but rather things that are specific to (the core component
  of) the node this config is on, like the node_nr
- `mConf.deployment` holds things that apply to the entire deployment,
  like number of nodes and so on
- `srvConf` refers to a service configuration

Still some work to be done in UI to adapt the config wizard, because
it's not split into a setup wizard (which works) and the old config
wizard (which needs work).