Fixes to resolve cold-boot and setup issues
The bundles a bunch of changes to core/api/ui to address some issues
with cold-start or initial setup, such as waiting for containers to be
available, ensuring that files are present on the filesystem and so on.

Also adds logic to reconfigure both API and core dynamically without the
need to restart the container.