wip: Working on generating .deb packages
This is a bit of a work-in-progress to have Morio generate its own
client packages. I'm starting with .deb packages, and they are building
now, and I'm also adding a page to the UI where people can generate
them.

This will need to be cleaned up and expanded to also generate other
formats. However, since things are building now, I wanted to commit my
work so far.