feat: Added downloads page and api endpoints
Since we're allowing people to build/create client packages, we should
also have a way to inform people what sort of pacakges are
available/have been created before.

So I've added and API endpoint to get the list of files, as well as the
ability to download static files from the API (this way we don't have to
spin up another service to serve static files as Traefik does support
serving files.

Then, I added a page in the UI to see the downloads too. Here however,
rather than just dump a file list, we present a nice view with the
various available client packages to download.

Also fixed the fact that the revision was not updated when a new client
package was built.