[ui] Added terminology pages and components
There's a bunch of terminology we use and I feel that over time that
might become a bit overwhelming for new users.
Since we now include the documentation inside Morio, I wanted a system
where we can define terms/terminolgy in one place and then use that
shared definition to provide help where it's used.

This implements that first steps of that. Works only in MDX for now, but
that might be the only place we need it.