[core/api] Split config from settings
This draws a line between settings (what the user provides) and config
(what core generates).

We were taking the user input as config and then extending it with
generated config, which was not very intuitive when the user would then
want to edit their settings since now things look so dramatically
different.

So now, user settings are kept mostly as-is, and configuration is
generated from it. The idea being that this is more intuitive:

- Settings: Provided by the user
- Config: Generated by Morio Core