[ui] wip: More refactoring forms to deal with setup/settings
At some point, we may need to refactor this and use something like
react-hook-form, but for now we're trying to get an MVP up and running
so I'm not going to get sidetracked by that now.

This essentially switches to managed forms that update the mSettings
state object directly. However, there's an escape hatch for things like
connector inputs/outputs where you can manage the state locally, and
then only at the end click a button to update the global mSettings
state.

This alloys you to configure the input, and only when you have all the
data, update the settings with it.
This is required because the input/output ID under which the settings
are to be anchored is not yet known when we render the form.