[ui] Added support for tokens
This adds support for tokens.vars and tokens.secrets.

The latter being the driving factor behind these changes as we need a
way to keep sensitive data in the settings (like passwords) without
storing them as plain text.

So secrets do that, and will be encrypted at rest. The variables just
reuse the same logic, minus the encryption.

Also handles the UI aspects of this.