feat: Initial implementation of authentication/RBAC in morio
This needs to be documented, but this adds authentication for all
HTTP-based access in Morio (only for direct access to the Kafka API do
we use mTLS).

The idea here is to support different authentication providers, of which
the Morio root token (mrt) is the only one that is active
out-of-the-box.

The way it works is that we use Traefik middleware to forward all
requests to a central authentication endpoint (`/auth` on the management
API). This endpoint will then check the access.

Note that this endpoint does not use the prefix used for routing in
Traefik so it is only accessible from inside the Docker network, and not
exposed to the outside world.

For authentication as users, we expose the ${prefix}/login endpoint.
This will return a JSON web token (JWT). The frontend will store this in
a `morio` cookie, which is one of the ways that we support
authentication.

We will (later) add support for passing the token as a Bearer token in
the headers too.