wip: Working towards supporting IPDs
Started adding support for different authentication/identity providers.
Both in the UI where we add them in the settings, as in api/core where
authentication is not handled yet, but we already support listing the
IDPs that are setup to dynamically load then on the login screen.

Next step: Actually implement auth. We'll start with LDAP as Active
Directory is a common IDP.