[core] fix: Issues with lifecycle container configuration
The container configuration was generated before the preCreate
lifecycle hook. As a result, the extra container labels that were added
to the service config were not on the container (since its configuration
was already generated based on the earlier configuration without the
extra labels).

This fixes that by postponing the generation of the container
configuration until after the preCreate lifecycle hook so that any
changes made to the service configuration by that hook are taken into
account.

In practical terms, the labels added are for Traefik's ACME
configuration.

This also fixes an issue introduced when refactoring the lifecycle
defaults where we were not passing in the service name.