[api] Implement RPKV client on top of KafkaJS
To add support for local user accounts in Morio, we need a place to
store that account data. We could write them to disk, but that
complicates matters when we run a clustered Morio deployment since now
we have to keep this data in-sync somehow, which is a challenge in
distributed systems.

A distributed database would be the best solution for that challenge,
but adding one would increase the complexity of the Morio setup.

But, we already have the RedPanda brokers who are a distributed storage
solution. Granted, using them as a database is not the most typical
use-case, but it can be done.

This implements an RPKV client (RedPanda Key Value client) that
implements the 'rpkd.set()' and  'rpkv.get()' methods that set and get a
key from the kv store respectively.

How it all works under the hood is out of scope, but it does.
Futhermore, I've implemented helper methods for accounts on top of this
such as loadAccount(), saveAccount(), and storeLastLoginTime().