[core] fix: Ensure ca certificate is always available
This adds the reload lifecycle hook giving services a way to do
housekeeping whenever core reloads.
In this case, it loads the ca's root certificate from disk and adds it
to the store. Something that needs to happen when core reloads, even
when the CA is running and requires no changes.