[api] fix: Made core loading at bootstrap more robust
When API reloads at a time core is not availble (because perhaps it's
also reloading) the fetching of config & info from core will fail, and
the API will fail to properly bootstrap.

While the call was wrapped in an `attempt()` call, so it should in
principle have another go at it when it fails, attempt only checks to
see whether the return value is truthy. But a fetch from core via the
client always returns an array, so it's considered a success even when
it failed.

This extends attempt() so you can pass in a custom method to validate
the result, which we do to ensure the config is loaded properly from
core.

While we're at it, we apply the same logic to also load the info
endpoint from core.