chore: Fix production build issues, add version script
There was a few things that were working ok in the development build,
but were not ok in a production build.

Production builds of core & api now also use the pm2 process manager
inside the container. This allows us to restart/reload without having to
restart the container, as well as a faster restart after a crash.
It also allows us to cap the memory use and trigger a restart when
memory grows.

There's also the new `npm run reversion` script to deal with versions as
going forward we will put out new versions from time to time to test the
upgrade process.