wip: Work towards stand-alone install
Depite our best efforts, in some places the code is relying on the git
repo structure being present.
That is fine in development mode, but we need to avoid this in
production as when Morio is deployed it is done on a system that does
not have the repo available.

These are a bunch  of changes that solve some of the dependency issues,
in particular how we volume-map host OS folders into the containers.