[core] fix: Always recreate containers after initial deploy
This bug was caused by attempting to optimize when to recreate
containers. Initially, we would always recreate the container when
reconfiguring core. However, often that is not needed so we put checks
in place to only recreate when required.

Problem is that when Morio runs in ephemeral mode, there is no TLS
configuration on traffic. When you deploy initial settings, core will
reconfigure itself, but because nothing has really changes for
containers like ui, it will not be recreated.

Since the labels required to setup Traefik TLS are added in the
preCreate lifecycle hook, they are never added. So even though the CA
is up and available, Traefik will not use ACME and just fall back on its
default certificate.

This fixes that by allowing to pass hookProps into the reconfigure
method as well as the default preCreate hook. We can then use that data
to make a better/more informed decision on whether to recreate the
container or not.