[core] fix: Wait before requesting broker certificate
We were already waiting for the CA to come up before requesting a
certificate for the broker. However, this created a race condition where
sometimes the JWT timestamp was older than the epoch/start of the CA
lifetime.

To avoid that, we now wait for the CA to come up, and then wait an
additional 2 seconds before requesting the broker certificate.