chore: Avoid shrinkwrap in Docker build
While `npm ci` is faster than `npm i` it does require the lock file to
be in sync with the dependencies. That can become non-trivial in a
monorepo as dependencies can be hoisted to the workspace root not to
mention that the Docker build context is limited to the subfolder.

So, to make life easier on ourselves we use npm install for production
builds.