[core] fix: Write client ad-hoc files to source, not destination
The core container ships with the base client files.
Those files are copied by the dbuilder to build the client package.

The previous behavior, the ad-hoc client files (certificates and so on)
were copied to the client destination, only to be overwritten again when
dbuilder copies the source files over the destination.

So now core creates them in the client source files, from where they
will be copied to the destination by dbuilder.