[ui] chore: Logout user when in ephemeral mode
This is a bit of a nich use case, but when running in development mode
and you recreate a morio instance, the user will still be logged in with
the old token when you connect to the (ephemeral) instance.

You won't be able to do anything, but the token being there is perceived
by the UI as everything being ok.

Now that there's a whoami endpoint, we connect to that and if the
instance in in epehemeral mode (which this endpoint will return) then
we always log out the user.