[core] feat: Do not recreate API container during tests
This carves out an exception to not restart the API service while
running tests.

The reason the API needs to be restarted is so that we can add the
labels for the Traefik TLS configuration, which can only be done when
the container is created.

However, the tests run inside the API container and configure to
localhost over HTTP (they do not go through the proxy) so there is no
need to restart the API in this case.