[core] wip: Work on running core without root privileges
This is a work in progress to see if we can run Morio's core container
without root privileges. Doing so requires some work because we need
elevated privileges to:

- Talk to the Docker daemon
- Write files in places regular users don't have access to
- Changes permissions/ownership of files

This commit brings this effort as far as running Morio in ephemeral mode
as a regular user. To make that happen, a few things are required:

- On the host OS, the docker socket needs to have a group that allows
  access to it
- Inside the core container, the morio user needs to be added with a
  group with the same GID as on the host OS
- There's a highly detailed sudoers configuration that details what
  morio can do with elevated privileges

The biggest problem is matching the host os GID to the one inside the
container, since this needs to be done when building the container.
As such, this would add a lot of complexity to the initial Morio setup
if people either have to create a specific group with a specific GID and
then configure access to the Docker socket or -- even worse -- have to
build their own Morio core image.

However, when we consider the 'appliance' model for deployment where we
control both host OS and the container images, then this is trivial to
implement.

As such, I am currently leaning towards making this opt-in and add some
abstraction that would allow Morio to work in both modes.

However, before doing more work on this, I want to mull it over a bit
and talk to some colleagues about it, so I'm parking this in a feature
branch for now.