[api] fix: Prevent RPKV lookups from hanging when topic is empty
We use RedPanda as a key/value store to stare account info, something we
refer to internally as RPKV.

We have various methods to search/get records, but they hang when there
is no data whatsover. This fixes that by keeping track of the number of
fetches that occured, and the number of records we found.

Since the very first fetch event is raised before the very first message
event, we check on the second fetch whether there we have seen any
records, and if we did not, we terminate the consumer.

This still required a timeout to expire twice, so that makes things slow
when there is no data. But this should be a rather niche case.