feat: Out with rdpv, in with db
Up until now, we had resisted the temptation to add a database service
to the Morio deployment, instead opting to store account and api key
data in the RedPanda broker, using a compacting topic and some custom
code to make a simple KV store out of RedPanda (aka rpkv).

However, when running unit tests, it becomes clear that this can be slow
at times. That's because looking up a value requires us to setup a Kafka
consumer, then connect to the broker, subscribe to a topic and then
fetch all records from the start.
This can cause a noticeable delay because while streaming is fast, the
initial setup is not so fast.

After mulling this over during the weekend, I decided that it would be
better to bite the bullet now and add a database service, rather than
having to re-visit this topic later. Problem is: What database. I really
don't want to add the complexity and resource hog that would be
something like Postgresql, instead I which I could just use sqlite but
that will cause problems down the road when we have to support
clustering?

Or will it? I did some tests with rqlite this weekend and it seems to be
everything we need and not more. So this rips out rpkv in favor of a
database service. This service is available to a REST API that is not
exposed, so it's considered an internal service.