feat: implemented initial support for clustering (#6)
This is far from ready, but contains the initial code for Morio to setup itself as a clustered deployment.
Specifically, this brings up 3-node (well, should work with more too, but was tested with 3) a 3-node Docker Swarm cluster.

It does not do anything apart from that (yet) and it breaks in many places. But it's something.