feat: Initial support for clustering
This adds (initial) support for clustering. In other words, Morio deployments where multiple nodes for a distributed system that is highly available.

I started this in the `clustering` branch and my initial plan was to rely on **Docker Swarm** to provide an overlay network that would encrypt communication between nodes. That way, we could keep the services simple as they can just talk to each other in plain text.

However, while working on that, I found the swarm overlay network to really be rather unreliable and what should have been something that _just works_ was more something that _just doens't work_.
In addition, the handling of the two different scenarios (local containers, and swarm services) with their subtle differences also added a lot of complexity to the code base.

So, at some point I decided that this was the wrong path, and instead ripped out all Swarm code, and switched to local containers again thereby making things a lot more simple.
The _downside_ of this is that we now no longer have encrypted networking between cluster nodes, and thus need to handle this ourselves. However, I do feel that ultimately that is the better option anyway, and ensuring encrypted connections between cluster nodes/services is arguably easier to debug/troubleshoot than relying on a service mesh that is fragile. One could also argue that it is _the zero trust way_.

I am certain there's going to be bugs here and unforeseen corner cases. But for what it's worth, things I've tested:
- Deploy a standalone Morio node
- Deploy a 3-node Morio cluster
- Bring down a node in the cluster, then bring it back up, cluster comes back up as expected, services on other nodes keep working

There are certain things that remain to be done. Known issues let's say:

- The provisioning of certificates in Traefik
- Access control for HTTP and KAFKA access
- Adapting the UI to all the latest changes

However, I felt it best to merge this feature branch at this state and start documenting things so that it would be easier for others to get involved if they would like to.