chore: Updating GitHub pipelines to run the CI/CD we want (#22)
**Clean up GitHub pipelines**

GitHub sets up a bunch of pipelines that are not always what we want.
For example, it sets up a prettier pipeline that will commit changes to the repo inside the pipeline. That's an absolute nightmare because now each time to your feature branch that has a PR open, your local branch will have diverged from the remote and you have to rebase. Not to mention that those commits are pointless because we run prettier anyway.

There's also a pipeline for running eslint and prettier on changed files. This too is unneeded because we do that in the pre-commit hook, and we run prettier/eslint on all files anyway.

**Bump pipeline NodeJS version to 20**

Also bumped the NodeJS version in the pipelines to lts/iron (node 20) since that was released for GA a few weeks ago.
I will also ensure we migrate to Node 20 everywhere else, although that's probably not all going to go in this PR.

**Linting of more files**

There were some files that weren't properly linted yet. Typically in either `docs` or `ui`. This PR also fixes that.

**Fix UI build**

After the cluster changes, the UI would no longer build due to some errors. I've fixed them for now. It will need a bit more polish, but I wanted the build to pass so we could build/publish the container image.

Which brings me to...

**Build container images in pipelines**

Last but not least, I've added pipelines to automatically build the container images (currently only, api, core, dbuilder, and ui are added because the rest of the services haven't been created yet).

This will not only build the images, it will also push them to the Docker registry under the [itsmorio](https://hub.docker.com/u/itsmorio) namespace.

Normally, we would build these only on a new release. However, for now they build on pull request so that I could test it, and that other could see the results of the new pipelines in this PR.

So, I intent to merge this as-is, and then update the pipelines again to run on release, and finally create a (GitHub) release which should trigger them again.