feat: Added watcher service
This adds a new Morio service: Watcher
Under the hood, it is Elastic heartbeat, so this is for running checks,
or watching for trouble if you will.

I've setup internal monitors for this for now, although we will have
to re-visit this as it probably does not make sense to run a watcher
instance on every node in clustered deployments.

This also includes a fix where the RedPanda broker was only trusting
the root CA certificate, rather than the intermediate CA certificate.