feat: Make Morio dev work on MacOS
My plane was delayed for over an hour so I got curious what it would
take to make the Morio development environment work on MacOs.

Turns out, not so much. Just a way to override the Docker log driver.
Then, once I boarded the plane I realized that I cannot make it work if
I do not have a routable IP address, so then I added a fix for that too.

This is a small change, but now I can actually work on the plane which
is nice :)