wip: Use SASL for authorization until mTLS works
To access the Kafka API, we use mTLS for authentication.
We also want to use it for authorization, but unfortunately, when
enabling authorization based on mTLS, the broker fails to extract the
principal (think username) from the X.509 certificate for reasons that
at this point are not clear to us.

Because this blocks us from adding authorization, we have for the moment
side-stepped the problem be reconfiguring the broker to use SASL for
authorization. This means that authorization works now, so we are not
blocked from releasing new updates.

This is supposed to be a temporary measure until we get authorization
based on mTLS to work as intended. Adding SASL does not only
complicate the configuration, it also complicates client configuration.
As a direct result, we have disabled the watcher service for now, since
logstash requires a JAAS config, as well as a Java keystore to be able
to talk to a SASL-enabled broker over TLS, and handling all that is
frankly a bunch of complexity for very little benefit.

So, for now we use SASL, we have created two users:
 - root, with password the same as the Morio Root Token, is a superuser
 - morio-client, with password the same as the cluster UUID can only
   produce data, not read data or manage the broker

Whether this will remain for a 0.4 release will depend on how much
progress we make with enabling mTLS based authorization, but for now
SASL will do.