fix [moriod]: Do not attach morionet in systemd
Do not attach the core container to morionet in the systemd unit file
for moriod as the network does not yet exists on intial setup.

Core will detect the absense of the morionet network and create it, after
which it will attach itself to it.
That mechanism was already in place, so this only updates the systemd unit
file which was not updated to not attach to morionet.