feat: Add support for building pre-release container images
This adds changes to tag containers as `next` when there is a `-`
character in the version number (eg: 0.4.1-rc.1). If not, they will be
tagged with `latest`.

This is required so we can release pre-release version to test without
running the risk that people will run these in production without
meaning to.