feat [core]: Allow preseeding of client templates
This extends the settings to allow preseeding of client templates.
This is an act in two parts. These changes allow the preseeding of these
templates. A future update will ensure that these templates are bundled
when a client packages is built. However, that work relies on #70 so I
will add it later.

This also includes a fix where the reseed action would only be allowed
if the base config was preseeded. Now that preseeding is a more generic
concept, this was adapted so that reseeding can always happen when there
are preseed.git settings.