feat [dbuilder]: Build amd64 and arm64 packages
This adds support for the Morio client on arm64.

The client itself supported Linux on arm64 since #70, but this adds
support to build Debian client packages within Morio for both the amd64
and arm64 architectures.

The packages will be hosted on the APT repo that is hosted by the Morio
collector, so things should _just work_ when installing the client via
the automated installer.