release-candidate: 0.5.0-rc.1 (#73)
This is a release candidate for 0.5.0. Here's what's changed since 0.4.2:

**Added**

- [api] Allow the export of key data
- [api] Allow building of client-repo packages
- [client] Ported the Morio client to Go
- [core] Allow the export of key data
- [core] Allow preseeding key data
- [core] Generate GPG key pair to sign packages
- [core] Automatically build a client packages on initial setup
- [core] Automatically build client installer packages on initial setup
- [core] Provide automated install script for clients
- [core] Support optional mTLS for HTTP behind the `ENFORCE_HTTP_MTLS` feature flag
- [core] Allow building of client-repo packages
- [core] Support preseeding of client templates
- [core] Added support for the RESEED_ON_RELOAD feature flag
- [core] Added support for Debian client packages for amd64 and arm64
- [dbuilder] Maintain APT repository of client packages
- [ui] Allow the export of key data
- [ui] Allow loading key data file in setup wizard
- [ui] Allow building of client-repo packages
- [watcher] We have enabled the watcher service which was disabled due to SASL
- [web] Added the web service to serve files over HTTP

**Changed**

- [core] Renamed the `HEADLESS_MORIO` feature flag to `DISABLE_SERVICE_UI`
- [proxy] Redirect HTTP to HTTPS

**Fixed**

- [api] Replace 'next' tag with 'testing'
- [core] Replace 'next' tag with 'testing'
- [core] Allow reseeding when the base config is not preseeded
- [broker] Fixed the issue with mTLS that was the reason for the SASL configuration as a temporary workaround
- [core] Guard against untagged images when filtering docker image list
- [dbuilder] Replace 'next' tag with 'testing'
- [moriod] Correct location of version file
- [ui] Replace 'next' tag with 'testing'
- [ui] Use correct spinner in setup wizard

**Removed**

- [broker] We have removed SASL since we now use mTLS based authentication and authorization

**Security**

- [broker] We now use mTLS based authorization (ACL)
- [core] Make console a broker superuser now that an ACL is enforced
- [core] Configure broker ACL for watcher service
- [core] Store the salted&hashed the root token. Earlier, it was stored in clear on disk to facilitate development, no longer.
- [core] Support optional mTLS for HTTP behind the `ENFORCE_HTTP_MTLS` feature flag

**Notes:**

* feat [core]: New RESEED_ON_RELOAD feature flag

Enable this flag to reseed Morio when reloading.

This flag comes into play when you have have preseed.git settings that
make Morio clone a git repository locally. This gives you various ways
to keep your configuration under version control, as outlined in the
Preseeding Guide.

By default, Morio will not re-clone the repository when you make a
configuration change, or when you restart Morio. Instead, it will only
update the git repository from the remote when you explicitly reseed
Morio through the reseed API endpoint, (or via the UI which uses this
endpoint under the hood).

When you enable this flag, Morio will reseed (thus update the local git
content from the remote) whenever you restart it, or update it’s
configuration.

* feat [dbuilder]: Build amd64 and arm64 packages

This adds support for the Morio client on arm64.

The client itself supported Linux on arm64 since #70, but this adds
support to build Debian client packages within Morio for both the amd64
and arm64 architectures.

The packages will be hosted on the APT repo that is hosted by the Morio
collector, so things should _just work_ when installing the client via
the automated installer.

---------

Co-authored-by @serge-tellene