chore: Move to single lodash import
We used individually packages lodash methods, but those seem to lag in
version upgrades. This caused vulnerability warnings on the repository
which may very well be false positives, but it's not a good look.

This changes the individual dependencies to the single lodash dependency
from which we then load the specific methods we need. If all goes well
the rest will be tree-shaked out of the bundle and it won't make much of
a difference.