chore: Update dependencies (#74)
* chore: Move to single lodash import

We used individually packages lodash methods, but those seem to lag in
version upgrades. This caused vulnerability warnings on the repository
which may very well be false positives, but it's not a good look.

This changes the individual dependencies to the single lodash dependency
from which we then load the specific methods we need. If all goes well
the rest will be tree-shaked out of the bundle and it won't make much of
a difference.

* chore: Updated container dependencies

This upgrades the versions of various Morio services:

- [broker] Upgraded RedPanda rom v24.2.5 to v24.2.7
- [ca] Upgraded Step CA from v0.27.4 to v0.27.5
- [connector] Upgraded Logstash from v8.15.1 to v8.15.3
- [db] Upgraded Rqlite from v8.30.5 to v8.32.3
- [proxy] Upgraded Traefik from v3.1.4 to v3.1.6
- [watcher] Upgraded Heartbeat from v8.15.1 to v8.15.3

* Review

---------

Co-authored-by: @serge-tellene