fix [api]: Token renew endpoint now uses correct passphrase
This is a holdout from the earlier changes we made where we no longer
use the Morio Root Token as the passphrase for Morio private key, in
anticipation of hte ability to rotate the root token.

However, in this particular API endpoint, we were still passing in the
root token to unlock they key, rather than the new unseal attribute.