wip: Work on HTTP output settings UI for connector
There was already a placeholder for the HTTP output in the connector
settings UI, but it did not actually do anything.

This is the start of populating the form with the many many different
options that are available for the HTTP output.

This also adds a few new smaller features such as the ability to specify
a help link which will automatically be included with a little '?' icon.