chore: Do not tie apt repos to a Debian release
Our APT repositories were specifically tied to Debian 12 (bookwork).
However, there's not reason for this, as they will work just the same on
a different Debian version, or Ubuntu or another distro using apt.

So we have updated our infrastructure to ensure our repositories are
version agnostic.

In addition, this includes changes to how the packages are build that
make them more discoverable for new contributors. Specifically, apt
source files are not included in the repo, rather than generated
dynamically with a script.