[core] fix: Improve robustness of preseeding
This addresses two issues:

- Avoid writing settings to disk when preseeding fails
- Handle tokens in preseed settings

The first problem happened when the preseed settings where valid
according to the schema (so they pass validation) but invalid in
practice (for example, failed authentication to the git provider).
In such a case, the reseedHandler would return `false` instead of
the preseeded settings, but we did not check this, and just wrote
`false` to disk as the new settings. This bricked the Morio setup.

Now, when preseeding fails, we back out and do not process the request
to update the settings, and return an error instead. This should make
to harder to brick your Morio setup.

The second issue has to do with token substitution in the preseed
settings. If you're using a secret or a var in the preseed settings,
they were not injected. Now they are, and we also inject tokens from
both the running setting, as well as the new ones.