feat: Flanking services configuration in UI and orchestration (#223)
This is a first batch of changes to improve the way flanking services are handled.
This focuses on the settings that allow users to change where flanking services should run (changes in UI and the settings schema), as well as the changes required in core to ensure orchestration enforces the desired state of the various services.

This lacks implementation of cross-cluster access to flanking services. For example:
- The API assumes that a DB service is available locally, but that is not the case on a flanking node
- The Tap service assumes that the Cache service is available locally, but that is not the case on a flanking node

A future update will address this by ensuring cross-cluster access to these services secured with mTLS, but I wanted to keep this pull request from getting too large.

**Added**

- [ui] Add UI to manage flanking nodes and services
- [core] Add selective deployment of flanking services

**Changed**

- Allow users to define where to run flanking services
- [api] Run in PM2 and restart when memory exceeds 250MB
- [core] Run in PM2 and restart when memory exceeds 250MB
- [core] Cluster join/heartbeat changes for flanking nodes

**Fixed**

- [api] Fix an issue in IP resolver validation
- [api] Fix an issue in flanking node validation

---------

Co-authored-by: serge-tellene <114212556+serge-tellene@users.noreply.github.com>