[core] chore: Add .internal network alias to containers
This is a matter of convenience to make it easier to configure Morio in
an environment that requires a proxy to access the internet.

When core needs to preseed settings, it can be setup with a proxy by
changing the sytemd unit file. However, once a proxy is configured, we
still need to avoid that inter-container connections go via the proxy.

Here, we add an [name].internal network alias to all containers and we
use this when connecting internally (eg: core would connect to
api.internal). This way, you can add '.internal' to your NO_PROXY
environment variable and things should _just work_.