[config] fix: Incorrect volume location for rqlite data
The volume path points to the standard rqlite data location
(/rqlite/file) but core will override this to store data in /data

As a result, data was not stored outside the container, so when core
recreates it (after it is destroyed by stopping moriod) all data is
lost.

This fixes that.