[tap] feat: Morio granular loading of tap processors (#262)
This refactors the tap service to allow more fine-grained control when dynamically loading stream processing code.
This goes along with flanking changes in Moriohub, as stream processors now have a different signature and export.

At the high level, a stream processor can subscribe to 3 levels:

    topic: This must be set, as it is the Kafka topic to subscribe to
    module: This is the Morio module. If it is unset, or * we will subscribe to all modules
    dataset: This is a further subdivision. For example, for metrics it holds the metricset, for logs, the logset and so on. If not, set or the value holds * we'll subscribe to all datasets.

In addition, dynamicly loaded processors should be placed in a top-level folder that has a name that is meaningful for its source. For example, processors from moriohub are nested under the moriohub folder.
This allows people to selectively overwrite files that are included.

We will document this in a stream processing guide when things stabilize.